project/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
blinker==1.9.0
blis==0.7.11
cachetools==6.2.0
camelot-py==1.0.9
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.1
cymem==2.0.11
distro==1.9.0
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.19.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.1
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain-core==0.3.76
langchain-openai==0.3.33
langcodes==3.5.0
langsmith==0.4.31
language_data==1.3.0
lazy_loader==0.4
marisa-trie==1.3.1
MarkupSafe==3.0.2
mpmath==1.3.0
murmurhash==1.0.13
narwhals==2.5.0
networkx==3.5
ninja==1.13.0
nltk==3.9.1
numpy==1.26.4
openai==1.109.1
opencv-python==4.9.0.80
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
orjson==3.11.3
packaging==23.2
pandas==2.3.2
pdf2image==1.17.0
pdfminer.six==20250506
pillow==11.3.0
portalocker==3.2.0
preshed==3.0.10
protobuf==6.32.1
pyarrow==21.0.0
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pypdf==5.9.0
pypdfium2==4.30.0
pytesseract==0.3.10
python-bidi==0.6.6
python-dateutil==2.9.0
python-dotenv==1.0.0
pytz==2025.2
pywin32==311
PyYAML==6.0.3
qdrant-client==1.15.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.3.2
scipy==1.16.2
sentence-transformers==5.1.1
sentencepiece==0.2.1
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
smart-open==6.4.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
streamlit==1.50.0
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.6.0
tifffile==2025.9.20
tiktoken==0.11.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.2
typer==0.9.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.3.4
wheel==0.45.1
zstandard==0.25.0
